fix(38): fold F24+F25 — one owner for the contract's tables and receipt grammar
