+++ docs/features/38-workflow-status-sensor-script/review-findings.md
@@ -39,8 +39,8 @@ symlink gap the confinement leaves open).
 | F21 | scripts/workflow-status.mjs:105-121 (projectPath/readProject; same leaf-only pattern pre-execution-snapshot.mjs:167) | security | med | fix-now | fold into phase | yes |
 | F22 | docs/features/38-workflow-status-sensor-script/ACCEPTANCE.md (SPEC criterion A:24 row dropped) | workflow | med | fix-now | replan-in-unit (user-confirmed manifest amendment, then /review-plan 38) | no |
 | F23 | scripts/workflow-status.mjs:1064 | code | med | fix-now | fold into phase | yes |
-| F24 | scripts/workflow-status.mjs:586-593 | code | med | fix-now | fold into phase | no |
-| F25 | scripts/workflow-status.mjs:482-500 | code | med | fix-now | fold into phase | no |
+| F24 | scripts/workflow-status.mjs:586-593 | code | med | fix-now | fold into phase | yes |
+| F25 | scripts/workflow-status.mjs:482-500 | code | med | fix-now | fold into phase | yes |
 | F26 | docs/workflow/MIGRATION.md:12-13 + .es.md:12-13 + ORCHESTRATION.md:41-43 + .es.md:43-44 | usage-docs | med | fix-now | fold into phase | no |
 | VF-1 | docs/workflow/REPOSITORY_STATE.md:14 · reviewer review-change · HEAD 0a9b740d9a6195f4788aaffdc60725b3260e96d3 · recheck live run: sensor emitted repository_state status "draft" on this repo's frozen table-form ledger; colon-regex read at :127 misses `\| Status \|` rows (template/docs/workflow/REPOSITORY_STATE.md:12 same shape) | code | confirmed | finding-mark | n/a | n/a |
 | VF-2 | gh pr view 150/24 → MERGED · reviewer review-change · HEAD 0a9b740d9a6195f4788aaffdc60725b3260e96d3 · recheck live run: 25/33 dependencies.unmet are done units with MERGED PRs outside the 20-PR window; 12 spurious substrate blockers | code | confirmed | finding-mark | n/a | n/a |
